% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capteurs_chateaubourg-data.R
\docType{data}
\name{capteurs_chateaubourg}
\alias{capteurs_chateaubourg}
\title{Coordonnées géographiques des capteurs telraam de chateaubourg}
\format{
Un dataframe à 76 lignes et 3 variables:
lat : latitude des différents points formant les segments de chaque capteur
lon : longitude des différents points formant les segments de chaque capteur
oidn : nom du capteur correspondant
}
\source{
\href{https://telraam.net}{Telraam}
}
\usage{
data(capteurs_chateaubourg)
}
\description{
Coordonnées géographiques des capteurs telraam de chateaubourg
}
\examples{
\dontrun{
data(capteurs_chateaubourg)
}
}
\keyword{datasets}
