% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courbe_par_heure.R
\name{courbe_par_heure}
\alias{courbe_par_heure}
\title{Courbe de la fréquentation horaire selon les sens et de la vitesse v85}
\usage{
courbe_par_heure(donnees, capteur, periode, vacances, JF, SM)
}
\arguments{
\item{donnees}{donnees issues d'un import de données Telraam}

\item{capteur}{capteur d'interet}

\item{periode}{interval lubridate pour la période}

\item{vacances}{choix de garder ou non les vacances ("oui","non","exclusivement")}

\item{JF}{choix de garder ou non les jours fériés ("oui","non","exclusivement")}

\item{SM}{choix des jours de la semaine pour la période 1 (vecteur composé de nombre entre 1 et 7)}
}
\value{
Un graphique plotly de la fréquentation et la vitesse V85 en fonction des heures de la journée
}
\description{
Courbe de la fréquentation horaire selon les sens et de la vitesse v85
}
\examples{
\dontrun{

data(chateaubourg)

periode <- interval(ymd_hms("2021/01/01 00:00:00"),ymd_hms("2022/01/01 00:00:00"))

courbe_par_heure(donnees_non_nulles,"Burel",periode,"oui","oui",1:5)
}
}
