% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courbe_periode.R
\name{courbe_periode}
\alias{courbe_periode}
\title{Comparaison de deux périodes pour le même capteur}
\usage{
courbe_periode(
  donnees,
  capteur,
  sens,
  mobilite,
  periode1,
  periode2,
  vacance1,
  vacance2,
  JF1,
  JF2,
  SM1,
  SM2
)
}
\arguments{
\item{donnees}{donnees issues d'un import de données Telraam}

\item{capteur}{capteur d'interet}

\item{sens}{sens du capteur ("Rgt"/"Lft"/"Toute")}

\item{mobilite}{vecteur des type de mobilité d'interet (parmis "car","heavy","bike","pedestrian")}

\item{periode1}{interval lubridate pour la période 1}

\item{periode2}{interval lubridate pour la période 2}

\item{vacance1}{choix de garder ou non les vacances pour la période 1 ("oui","non","exclusivement")}

\item{vacance2}{choix de garder ou non les vacances pour la période 2 ("oui","non","exclusivement")}

\item{JF1}{choix de garder ou non les jours fériés pour la période 1 ("oui","non","exclusivement")}

\item{JF2}{choix de garder ou non les jours fériés pour la période 2 ("oui","non","exclusivement")}

\item{SM1}{choix des jours de la semaine pour la période 1 (vecteur composé de nombre entre 1 et 7)}

\item{SM2}{choix des jours de la semaine pour la période 2 (vecteur composé de nombre entre 1 et 7)}
}
\value{
un graphique composé des courbes des deux périodes et des resultats d'un test de wilcoxon
heure par heure
Ou un message explicitant l'absence de données
}
\description{
Trace les courbes de fréquentation horaire pour les deux périodes et teste
l'égalité des répartition entre les périodes
}
\examples{
\dontrun{

data(chateaubourg)

periode1 <- interval(ymd_hms("2021/01/01 00:00:00"),ymd_hms("2022/01/01 00:00:00"))
periode2 <- interval(ymd_hms("2022/01/01 00:00:00"),ymd_hms("2023/01/01 00:00:00"))

courbe_periode(chateaubourg,"Burel","Lft",c("car","heavy"),periode1,periode2,
               "oui","non","oui","non",1:5,1:5)
}
}
