% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courbe_vitesse.R
\name{courbe_vitesse}
\alias{courbe_vitesse}
\title{Graphique de la part d'usagers atteignant certaines vitesses selon la ciculation}
\usage{
courbe_vitesse(Donnees, Id, orientation)
}
\arguments{
\item{Donnees}{Tableau issue de l'import de donnée Telraam}

\item{Id}{Nom du cateur d'interet}

\item{orientation}{Choix de la direction du capteur ("Toute"/"Rgt"/"Lft")}
}
\value{
Graphique
}
\description{
Trace sur un graphique le pourcentages d'usagers arrivant a dépasser respectivement
10km/h, 20km/h, 30km/h et 40km/h en fonction du  nombres d'usagers par heure (voiture plus poids lourds)
}
\examples{
\dontrun{

liste_capteurs <- c("9000002156", "9000001844")

liste_noms <- c("Burel","RteVitré")

data("clef")

Donnee <- import(liste_capteurs, liste_noms, clef)$donnee

courbe_vitesse(Donne, "Burel","Toute")

}
}
