% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation_carte.R
\name{creation_carte}
\alias{creation_carte}
\title{création de carte permettant de visualiser la correlation}
\usage{
creation_carte(donnees, dfgeo, segment, heure)
}
\arguments{
\item{donnees}{donnees issues d'un import de données Telraam}

\item{dfgeo}{infos geographiques issues d'un import de données Telraam}

\item{segment}{non du capteur de référence}

\item{heure}{heure choisie pour l'analyse}
}
\value{
une carte leaflet ou apparait le capteur de référence et les autres capteurs avec les correlations
compare le trafic total entre chaque capteur
}
\description{
création de carte permettant de visualiser la correlation
}
\examples{
\dontrun{
data(chateaubourg)
data(capteurs_chateaubourg)

creation_carte(chateaubourg,capteurs_chateaubourg,"Burel",9)

}
}
