% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation_carte_non_traitee.R
\name{creation_carte_non_traitee}
\alias{creation_carte_non_traitee}
\title{Visualisation des corrélations avec un capteur de référence sur une carte
sans traitement de saisonalité donc à heure et jour fixe}
\usage{
creation_carte_non_traitee(
  donnees,
  dfgeo,
  segment,
  sens,
  liste_seg,
  liste_sens,
  heure,
  jour
)
}
\arguments{
\item{donnees}{donnees issues d'un import de données Telraam}

\item{dfgeo}{infos geographiques issues d'un import de données Telraam}

\item{segment}{non du capteur de référence}

\item{sens}{sens du capteur de référence}

\item{liste_seg}{liste de capteur à comparer}

\item{liste_sens}{liste des sens correspondant}

\item{heure}{heure choisie pour l'analyse}

\item{jour}{jour choisi pour l'analyse (nombre entre 1 et 7)}
}
\value{
une carte ou les segments apparaisse sous la forme de point dont la taille est
proportionnelle à la valeur de la corrélation.
}
\description{
Visualisation des corrélations avec un capteur de référence sur une carte
sans traitement de saisonalité donc à heure et jour fixe
}
\examples{
\dontrun{
data(chateaubourg)
data(capteurs_chateaubourg)

creation_carte_non_traitee(chateaubourg,capteurs_chateaubourg,"Burel","Rgt",
                           c("RteVitré","ParisArcEnCiel","RueGdDomaine","ParisMarché"),
                           c("Rgt","Lft","Rgt","Lft"),9,1)

}
}
