% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courbes_non_traitees.R
\name{creation_courbe}
\alias{creation_courbe}
\title{Title}
\usage{
creation_courbe(donnees, segment, sens, segment2, sens2, heure, jour)
}
\arguments{
\item{donnees}{donnees issues d'un import de données Telraam}

\item{segment}{Premier capteur}

\item{sens}{Choix du sens du premier capteur ("Rgt"/"Lft"/"Toute")}

\item{segment2}{Second capteur}

\item{sens2}{Choix du sens du second capteur ("Rgt"/"Lft"/"Toute")}

\item{heure}{heure choisie pour l'analyse}

\item{jour}{jour de la semaine, entre 1 et 7 (lundi à dimanche)}
}
\value{
Une liste composé du graph et du coefficient de corrélation
}
\description{
Title
}
\examples{
\dontrun{

liste_capteurs <- c("9000002156", "9000001844")

liste_noms <- c("Burel","RteVitré")

data("clef")

Donnees <- import(liste_capteurs, liste_noms, clef)$donnee

graph <- creation_courbe(Donnees,"Burel","Rgt","RteVitré","Rgt",9,5)$graph
graph

creation_courbe(Donnees,"Burel","Rgt","RteVitré","Rgt",9,5)$correlation

}

}
