% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation_table_heure.R
\name{creation_table_heure}
\alias{creation_table_heure}
\title{Tableau de correlation entre un capteur de référence et d'autres capteurs}
\usage{
creation_table_heure(
  donnees,
  segment,
  sens,
  liste_seg,
  liste_sens,
  liste_heure,
  jour
)
}
\arguments{
\item{donnees}{donnees issues d'un import de données Telraam}

\item{segment}{capteur de référence}

\item{sens}{sens du capteur de référence ("Rgt"/"Lft"/"Toute")}

\item{liste_seg}{vecteur de capteurs à comparer}

\item{liste_sens}{vecteur des sens de capteurs à comparer ("Rgt"/"Lft"/"Toute")}

\item{liste_heure}{une liste d'heure pour lesquels on veut la comparaison}

\item{jour}{un jour (nombre entre 1 et 7)}
}
\value{
Un tableau donc les lignes sont les capteurs à comparés et les colonnes les heures choisies,
les valeurs correspondent aux coefficient de correlations de Pearson entre le capteur de reference
et le capteur de la ligne pour le jour entré, à l'heure de la colonne
}
\description{
Tableau de correlation entre un capteur de référence et d'autres capteurs
}
\examples{
\dontrun{

data(chateaubourg)

creation_table_heure(chateaubourg,"Burel","Rgt",
                     c("RteVitré","ParisArcEnCiel","RueGdDomaine", "ParisMarché"),
                     c("Rgt","Lft","Rgt","Lft"),8:20,2) \%>\% View()

}
}
