% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation_table_jour.R
\name{creation_table_jour}
\alias{creation_table_jour}
\title{Tableau de correlation entre un capteur de référence et d'autres capteurs}
\usage{
creation_table_jour(donnees, segment, sens, liste_seg, liste_sens, heure)
}
\arguments{
\item{donnees}{donnees issues d'un import de données Telraam}

\item{segment}{capteur de référence}

\item{sens}{sens du capteur de référence ()}

\item{liste_seg}{vecteur de capteurs à comparer}

\item{liste_sens}{vecteur des sens de capteurs à comparer}

\item{heure}{heure choisie pour l'analyse}
}
\value{
Un tableau donc les lignes sont les capteurs à comparés et les jours (nombre entre 1 et 7),
les valeurs correspondent aux coefficient de correlations de Pearson entre le capteur de reference
et le capteur de la ligne pour l'heure entrée, au jour de la colonne
}
\description{
Tableau de correlation entre un capteur de référence et d'autres capteurs
}
\examples{
\dontrun{

data(chateaubourg)

creation_table_jour(chateaubourg,"Burel","Rgt",
                     c("RteVitré","ParisArcEnCiel","RueGdDomaine", "ParisMarché"),
                     c("Rgt","Lft","Rgt","Lft"),9) \%>\% View()

}
}
