% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decouper_periode.R
\name{decouper_periode}
\alias{decouper_periode}
\title{Découper une période en "tranches" de 3 mois}
\usage{
decouper_periode(date1, date2)
}
\arguments{
\item{date1}{Caractère. Date de début au format "aaaa-mm-jj hh:mm:ss"}

\item{date2}{Caractère. Date de fin au format "aaaa-mm-jj hh:mm:ss"}
}
\value{
Dataframe avec une ligne par période de 3 mois et deux colonnes "debut" et "fin"
}
\description{
Découper une période en "tranches" de 3 mois
}
\examples{
\dontrun{
date1 <- "2021-03-01 12:35:21"
date2 <- "2022-07-01 03:15:33"

test_df <- decouper_periode(date1, date2)
}

}
