% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desaisonnalite.R
\name{desaisonnalite}
\alias{desaisonnalite}
\title{Séparation d'un signal en 3 part ( tendance, cycle hebdomadaire et bruit statistique)}
\usage{
desaisonnalite(donnees, col, model)
}
\arguments{
\item{donnees}{Un dataframe avec une colonne "date" (format lubridate) et une colonne cible}

\item{col}{Nom de la colonne cible}

\item{model}{un modèle (multiplicatif ou additif) pour la séparation}
}
\value{
une liste de 3 vecteurs: la tendance, le cycle hebdomadaire et le bruit statistique
}
\description{
Séparation d'un signal en 3 part ( tendance, cycle hebdomadaire et bruit statistique)
}
\examples{
\dontrun{

liste_capteurs <- c("9000002156", "9000001844")

liste_noms <- c("Burel","RteVitré")

data("clef")

Donnees <- import(liste_capteurs, liste_noms, clef)$donnee

Burel_Vitre <- tableau_horaire(donnees_nn_plus,c("Burel","RteVitré"),9,"Lft","Lft")

Burel_traitement <- desaisonalite(Burel_Vitre,"Burel","add")

}

}
