% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desaisonnalite_tab.R
\name{desaisonnalite_tab}
\alias{desaisonnalite_tab}
\title{Extraction du bruit a partir d'un tableau issu de "donnees_horaires_groupes"}
\usage{
desaisonnalite_tab(donnees)
}
\arguments{
\item{donnees}{Un tableau issu de "donnees_horaires_groupes"}
}
\value{
Un tableau à "nombre de capteur+1" colonnes
- "date" : date des mesures
- non de chaque capteur_sens du capteur : bruit statistique correspondant
à chaque colonne du tableau entré
}
\description{
Extraction du bruit a partir d'un tableau issu de "donnees_horaires_groupes"
}
\examples{
\dontrun{

segs <- c("Burel","ParisMarché","ParisArcEnCiel","RteVitré","RueGdDomaine")
dir <- c("Rgt","Lft","Lft","Rgt","Rgt")

donnees_synchrone <- donnees_horaires_groupes(donnees_nn_plus,segs,9,dir)

donnees_a_analyser <- desaisonnalite_tab(donnees_synchrone)

mod <- lm(Burel_Rgt~., data = donnees_a_analyser[,-1])
summary(mod)

}

}
