% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees_horaire.R
\name{donnees_horaire}
\alias{donnees_horaire}
\title{Sélectionne les périodes commune à 2 capteurs}
\usage{
donnees_horaire(donnees, segments, heure, dirA, dirB)
}
\arguments{
\item{donnees}{tableau issu de l'import de données Telraam}

\item{segments}{vecteur des 2 capteurs considérés}

\item{heure}{l'heure choisie pour l'analyse}

\item{dirA}{Sens choisi pour le capteur 1 ("Rgt"/"Lft"/"Toute")}

\item{dirB}{Sens choisi pour le capteur 2 ("Rgt"/"Lft"/"Toute")}
}
\value{
Un tableau à 3 colonnes
\itemize{
\item "date" : date des mesures
\item non du premier capteur : VL + PL dans la direction choisie
correspondant aux dates communes (pour le premier capteur)
\item non du second capteur : VL + PL dans la direction choisie
correspondant aux dates communes (pour le second capteur)
}
}
\description{
A partir de choix de capteur, de sens et d'une heure, isoler les
donnees sur des dates communes pour deux capteurs (car+heavy)
Nécessaire pour la plus part des fonction de comparaisons de 2 capteurs
}
\examples{
\dontrun{

liste_capteurs <- c("9000002156", "9000001844")

liste_noms <- c("Burel","RteVitré")

data("clef")

Donnees <- import(liste_capteurs, liste_noms, clef)$donnee

tableau_horaire(Donnees,c("Burel","RteVitré"),9,"Lft","Lft")

}

}
