% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees_horaires_groupes.R
\name{donnees_horaires_groupes}
\alias{donnees_horaires_groupes}
\title{Jointure sur les dates sur listes de segments}
\usage{
donnees_horaires_groupes(donnees, segments, heure, dir)
}
\arguments{
\item{donnees}{tableau issu de l'import de données Telraam}

\item{segments}{vecteur des 2 capteurs}

\item{heure}{l'heure choisie pour l'analyse}

\item{dir}{vecteurs des directions des capteurs choisis}
}
\value{
Un tableau à "nombre de capteur+1" colonnes
- "date" : date des mesures
- non de chaque capteur_sens du capteur : VL + PL dans la direction choisie
correspondant aux dates communes avec les autres capteurs
}
\description{
A partir de choix de capteur, de sens et d'une heure, isoler les
donnees sur des dates communes pour une liste de capteurs (car+heavy)
}
\examples{
\dontrun{

segs <- c("Burel","ParisMarché","ParisArcEnCiel","RteVitré","RueGdDomaine")
dir <- c("Rgt","Lft","Lft","Rgt","Rgt")

donnees_synchrone <- donnees_horaires_groupes(donnees_nn_plus,segs,9,dir)

donnees_a_analyser <- desaisonnalite_tab(donnees_synchrone)

mod <- lm(Burel_Rgt~., data = donnees_a_analyser[,-1])
summary(mod)

}
}
