% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphe_seuil.R
\name{graphe_seuil}
\alias{graphe_seuil}
\title{Graphique de la part d'usagers atteignant certaines vitesses selon la circulation avec un seuil}
\usage{
graphe_seuil(Donnees, Id, orientation)
}
\arguments{
\item{Donnees}{Tableau issue de l'import de donnée Telraam}

\item{Id}{Nom du cateur d'interet}

\item{orientation}{Choix de la direction du capteur ("Toute"/"Rgt"/"Lft")}
}
\value{
Un dataframe à 5 colonnes correspondant à l'absisse et aux ordonnées des courbes
}
\description{
Trace sur un graphique le pourcentages d'usagers arrivant a dépasser respectivement
10km/h, 20km/h, 30km/h et 40km/h en fonction du  nombres d'usagers par heure (voiture plus poids lourds)
Affiche un seuil calculé à partir d'un test de Darling Erdos
}
\details{
Cette fonction s'appuit sur le test de Darling-Erdos du package CPAT:
Curtis Miller (2018). CPAT: Change Point Analysis Tests. R package version 0.1.0.
https://CRAN.R-project.org/package=CPAT
}
\examples{
\dontrun{

liste_capteurs <- c("9000002156", "9000001844")

liste_noms <- c("Burel","RteVitré")

data("clef")

Donnees <- import(liste_capteurs, liste_noms, clef)$donnee

graphe_seuil(Donnees, "Burel","Toute")

}
}
