% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jours_feries.R
\name{jours_feries}
\alias{jours_feries}
\title{Import des jours fériés sur 25 ans}
\usage{
jours_feries()
}
\value{
un vecteur de date au format lubridate
}
\description{
Import des jours fériés sur 25 ans
}
\examples{
\dontrun{
data(chateaubourg)
jours <- jours_feries()

donnees_filtrees <- selection_date(chateaubourg,vacance) \%>\%
                        .$donnees_correspondantes

summary(donnees_filtrees)

}
}
