% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quatre_tests.R
\name{quatre_tests}
\alias{quatre_tests}
\title{Calculs de seuils à partir de la détection dans la cassure de courbes à partir de 4 tests}
\usage{
quatre_tests(Donnees, Id, orientation)
}
\arguments{
\item{Donnees}{Tableau issue de l'import de donnée Telraam}

\item{Id}{Nom du cateur d'interet}

\item{orientation}{Choix de la direction du capteur ("Toute"/"Rgt"/"Lft")}
}
\value{
Un dataframe à 5 colonnes correspondant à l'absisse et aux ordonnées des courbes
}
\description{
Cette fonction s'appuit lourdement sur le package CPAT:
Curtis Miller (2018). CPAT: Change Point Analysis Tests. R package version 0.1.0.
https://CRAN.R-project.org/package=CPAT
}
\examples{
\dontrun{

liste_capteurs <- c("9000002156", "9000001844")

liste_noms <- c("Burel","RteVitré")

data("clef")

Donnees <- import(liste_capteurs, liste_noms, clef)$donnee

quatre_tests(Donnees, "Burel","Toute")

}
}
