% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_date.R
\name{selection_date}
\alias{selection_date}
\title{Filtrage sur l'appartenance à une liste de périodes d'un dataframe}
\usage{
selection_date(donnees, liste_interval)
}
\arguments{
\item{donnees}{un dataframe avec une colonne "date" (format lubridate)}

\item{liste_interval}{une liste d'intervals lubridate}
}
\value{
Le data frame des lignes correspondants aux intervals et un dataframe complementaire
sous la forme d'une liste a 2 elements: donnees_correspondantes et donnees_complementaires
}
\description{
Filtrage sur l'appartenance à une liste de périodes d'un dataframe
}
