% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_date2.R
\name{selection_date2}
\alias{selection_date2}
\title{Filtrage sur l'appartenance à une liste de dates d'un dataframe}
\usage{
selection_date2(donnees, liste_date)
}
\arguments{
\item{donnees}{un dataframe avec une colonne "date" (format lubridate)}

\item{liste_date}{une liste de dates lubridate}
}
\value{
Le data frame des lignes correspondants aux dates et un dataframe complementaire
sous la forme d'une liste a 2 elements: donnees_correspondantes et donnees_complementaires
}
\description{
Filtrage sur l'appartenance à une liste de dates d'un dataframe
}
\examples{
\dontrun{
data(chateaubourg)
jours <- jours_feries()

donnees_filtrees <- selection_date(chateaubourg,vacance) \%>\%
                        .$donnees_correspondantes

summary(donnees_filtrees)

}
}
