% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_date.R
\name{test_date}
\alias{test_date}
\title{Test d'appartenance d'une date a une liste d'intervals lubridate}
\usage{
test_date(date, liste_interval)
}
\arguments{
\item{date}{Date au format lubridate}

\item{liste_interval}{Liste d'intervals lubridate}
}
\value{
Un booléen indiquant si la date appartient ou pas à la liste d'interval
}
\description{
Test d'appartenance d'une date a une liste d'intervals lubridate
}
\examples{
\dontrun{
date <- ymd_hms("2021/01/01 00:00:00")
interval1 <- interval("2020/11/02 00:00:00","2021/03/01 00:00:00")
interval2 <- interval("2022/02/01 00:00:00","2021/04/01 00:00:00")
Liste_interval = c(interval1, interval2)
test_date(date,Liste_interval)
}
}
