% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import des données Telraam}
\usage{
import(liste_capteurs, liste_noms, clef)
}
\arguments{
\item{liste_capteurs}{vecteur des identifiants des capteurs}

\item{liste_noms}{vecteur des noms à donner aux identifiants}

\item{clef}{Votre clef d'accès à l'API}
}
\value{
une liste de 2 objets: geomet donnant accès aux information géographique sur le capteur
et donnee donnant accès au données horaire pour chaques capteurs
}
\description{
Import des données Telraam
}
\examples{

liste_capteurs <- c("9000002156", "9000001844")

liste_noms <- c("Burel","RteVitré")

data("clef")

Donnee <- import(liste_capteurs, liste_noms, clef)$donnee
head(Donnee)


}
